made double surfpack::atanh(double x) into an inline function in the surfpack.h file that if defs to use BMA's log formula if on windows and otherwise calls ::atanh() per WJB's suggestion hopefully this will not cause problems in C++11 or later compliant compilers which places atanh in std::
when compiling with -Wall I encountered a comparison of ints to unsigned or size_ts in KrigingModel.cpp and fixed those by either changing the index from int to size_t or in one or two cases with a static_cast

In DirectANNModel.cpp -Wall revealed that finalSum (capitalization may be wrong it was a local variable) was uninitialized and used right before a loop to sum a bunch of things so I initialized it to zero.






git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@625 5a19d2f1-2ada-482b-9066-51d89674df4a

+++ src/surfaces/KrigingModel.cpp
@@ -41,7 +41,7 @@ surfdata_to_nkm_surfdata(const SurfData& sd, nkm::SurfData& nkm_sd)
     
     for(unsigned f_index=0; f_index<f_size; ++f_index) {
       derY[f_index].resize(der_order(0,f_index)+1);
-      for(unsigned der_order_index=1; der_order_index<=der_order(0,f_index); ++der_order_index)
+      for(unsigned der_order_index=1; static_cast<int>(der_order_index)<=der_order(0,f_index); ++der_order_index)
 	derY[f_index][der_order_index].newSize(num_points,nkm::num_multi_dim_poly_coef(x_size,-der_order_index));
     }
   }
@@ -184,14 +184,14 @@ double KrigingModel::variance(const VecDbl& x) const
 VecDbl KrigingModel::gradient(const VecDbl& x) const
 {
   nkm::MtxDbl nkm_x(1, ndims);
-  for(int i=0; i<ndims; ++i)
+  for(size_t i=0; i<ndims; ++i)
     nkm_x(0, i) = x[i];
 
   nkm::MtxDbl nkm_d1y(1, ndims);
   nkmKrigingModel->evaluate_d1y(nkm_d1y, nkm_x);
 
   VecDbl d1y(ndims, 0.0); 
-  for(int i=0; i<ndims; ++i)
+  for(size_t i=0; i<ndims; ++i)
     d1y[i] = nkm_d1y(0,i);
 
   return d1y;
@@ -200,7 +200,7 @@ VecDbl KrigingModel::gradient(const VecDbl& x) const
 MtxDbl KrigingModel::hessian(const VecDbl& x) const
 {
   nkm::MtxDbl nkm_x(1, ndims);
-  for(int i=0; i<ndims; ++i)
+  for(size_t i=0; i<ndims; ++i)
     nkm_x(0, i) = x[i];
 
   int num_lower_elem=(ndims+1)*ndims/2;
@@ -209,10 +209,10 @@ MtxDbl KrigingModel::hessian(const VecDbl& x) const
 
   MtxDbl d2y(ndims, ndims, 0.0); 
   int k=0;
-  for(int j=0; j<ndims; ++j) {
+  for(size_t j=0; j<ndims; ++j) {
     d2y(j,j)=nkm_d2y(0,k);
     k++;
-    for(int i=j+1; i<ndims; ++i) {
+    for(size_t i=j+1; i<ndims; ++i) {
       d2y(i,j) = nkm_d2y(0,k);
       d2y(j,i) = d2y(i,j);
       k++;
